Reorganize opportunity user requests
