--- app/Http/Requests/Backend/OpportunityUser/RemoveOpportunityUserRequest.php
@@ -1,55 +1,0 @@
-<?php
-
-namespace SCCatalog\Http\Requests\Backend\OpportunityUser;
-
-use Illuminate\Validation\Rule;
-use Illuminate\Foundation\Http\FormRequest;
-
-class RemoveOpportunityUserRequest extends FormRequest
-{
-    /**
-     * Determine if the user is authorized to make this request.
-     *
-     * @return bool
-     */
-    public function authorize()
-    {
-        return $this->user()->can('remove project user') || $this->user()->isAdmin();
-    }
-
-    /**
-     * Get the validation rules that apply to the request.
-     *
-     * @return array
-     */
-    public function rules()
-    {
-        return [
-            //
-        ];
-    }
-
-    /**
-     * Get the validation attributes that apply to the request.
-     *
-     * @return array
-     */
-    public function attributes()
-    {
-        return [
-            //
-        ];
-    }
-
-    /**
-     * Get the validation messages that apply to the request.
-     *
-     * @return array
-     */
-    public function messages()
-    {
-        return [
-            //
-        ];
-    }
-}
